+++ crates/mp/engine/qcommon/src/vm_fns.rs
@@ -923,7 +923,8 @@ pub fn VM_Create(
 
         // the stack is implicitly at the end of the image
         (*vm).programStack = (*vm).dataMask + 1;
-        (*vm).stackBottom = (*vm).programStack - crate::vm::vm_stack_consts::STACK_SIZE;
+        (*vm).stackBottom =
+            (*vm).programStack - crate::vm::vm_stack_consts::STACK_SIZE as i32;
 
         vm
     }
